Dataset for the program
This file has dataset for the program.
Please change the url before running the code. I have tried to create a function to check datatypes but couldnt do it. Still working on that part. Otherwise the program is working. It creates an output file in CSV format with additional columns. The output file can be opened in excel.